+++ src/interfaces/controller/common/guards/jwt.strategy.ts
